jwt testé + manque php

+++ api/data/main.py
@@ -17,19 +17,6 @@ def get_db():
         db.close()
 
 
-def is_admin(func):
-    def wrapper(*args, **kwargs):
-        payload = kwargs.get("payload")
-        if payload is None or "statut" not in payload or payload["statut"] != "admin":
-            raise HTTPException(
-                status_code=403,
-                detail="Vous ne possédez pas les droits nécessaires pour cette action",
-            )
-        return func(*args, **kwargs)
-
-    return wrapper
-
-
 @app.get("/produits", response_model=list[schemas.Produit])
 def get_produits(db: Session = Depends(get_db)):
     return [schemas.Produit.from_orm(pdt) for pdt in crud.get_produits(db)]
@@ -44,23 +31,31 @@ def get_produit(idP: int, db: Session = Depends(get_db)):
 
 
 @app.post("/produits", response_model=schemas.Produit)
-@is_admin
 def create_produit(
     produit: schemas.Produit,
     db: Session = Depends(get_db),
     payload=Depends(verifierTokenAccess),
 ):
+    if payload is None or "statut" not in payload or payload["statut"] != "admin":
+        raise HTTPException(
+            status_code=403,
+            detail="Vous ne possédez pas les droits nécessaires pour cette action",
+        )
     db_product = crud.create_produit(db, produit=produit)
     return schemas.Produit.from_orm(db_product)
 
 
 @app.put("/produits", response_model=schemas.Produit)
-@is_admin
 def update_produit(
     produit: schemas.Produit,
     db: Session = Depends(get_db),
     payload=Depends(verifierTokenAccess),
 ):
+    if payload is None or "statut" not in payload or payload["statut"] != "admin":
+        raise HTTPException(
+            status_code=403,
+            detail="Vous ne possédez pas les droits nécessaires pour cette action",
+        )
     db_product = crud.update_produit(db, produit=produit)
     if not db_product:
         raise HTTPException(status_code=404, detail="Produit inconnu")
@@ -68,12 +63,16 @@ def update_produit(
 
 
 @app.delete("/produits/{idP}")
-@is_admin
 def delete_produit(
     idP: int,
     db: Session = Depends(get_db),
     payload=Depends(verifierTokenAccess),
 ):
+    if payload is None or "statut" not in payload or payload["statut"] != "admin":
+        raise HTTPException(
+            status_code=403,
+            detail="Vous ne possédez pas les droits nécessaires pour cette action",
+        )
     db_product = crud.delete_produit(db, idP)
     if not db_product:
         raise HTTPException(status_code=404, detail="Produit inconnu")
